requirements.txt
requests
transmission-fluid
feedparser
bencoder